# perf(plugin-pnp): devirtualize `Locator` for package data cache