**What's the problem this PR addresses?**

When a `Package` has multiple virtual instances the `PnpInstaller` will fetch the same data about all of them instead of fetching it once and reusing it.

**How did you fix it?**

Devirtualize the `Locator` and use its `LocatorHash` instead of the virtual one

**Result**

Tested on the Storybook repo using PnP
```
YARN_IGNORE_PATH=1 hyperfine -w 1 --prepare="rm -rf .yarn/install-state.gz" \
                                                    "node ./master.cjs --mode skip-build" \
                                                    "node ./after.cjs --mode skip-build"
Benchmark #1: node ./master.cjs --mode skip-build
  Time (mean ± σ):      6.724 s ±  0.108 s    [User: 8.198 s, System: 1.216 s]
  Range (min … max):    6.594 s …  6.981 s    10 runs

Benchmark #2: node ./after.cjs --mode skip-build
  Time (mean ± σ):      6.328 s ±  0.079 s    [User: 7.864 s, System: 1.085 s]
  Range (min … max):    6.239 s …  6.523 s    10 runs

Summary
  'node ./after.cjs --mode skip-build' ran
    1.06 ± 0.02 times faster than 'node ./master.cjs --mode skip-build'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.